requirements.txt
streamlit>=1.32.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
yfinance==0.2.40
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pytz>=2024.1
scipy>=1.11.0
ta==0.11.0
requests>=2.31.0
python-dateutil>=2.9.0
multitasking>=0.0.11
lxml>=4.9.0
